requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
quart==0.19.4
asyncio==3.4.3
